# docs/openapi.full.yaml
openapi: 3.0.3
info:
  title: Biometric Face Attendance API (Full)
  version: 1.1.0
  description: |
    Exhaustive endpoint specification for the backend route surface.
    Includes all discovered routes from auth, admin, lecturer, student, and recognition blueprints.
  x-last-reviewed: 2026-04-14
servers:
  - url: http://localhost:5000
    description: Local development
...
components:
  securitySchemes:
    jwtCookie:
      type: apiKey
      in: cookie
      name: access_token_cookie
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    GenericObject:
      type: object
      additionalProperties: true
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
    StartSessionRequest:
      type: object
      required: [paper_id]
      properties:
        paper_id:
          type: string
    RecognizeFrameRequest:
      type: object
      required: [session_id, frame]
      properties:
        session_id:
          type: string
        frame:
          type: string
          description: data URL base64 image
...
    CreateStudentRequest:
      type: object
      required: [name, email, course_id]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        course_id:
          type: string
        department:
          type: string
        enrollment_year:
          type: integer
        mobile_no:
          type: string
        enrolled_papers:
          type: array
          items:
            type: string
...
    BulkPromoteRequest:
      type: object
      required: [student_ids]
      properties:
        student_ids:
          type: array
          items:
            type: string
        from_semester:
          type: integer
...
    AuthUser:
      type: object
      required: [_id, name, email, role, must_change_password]
      properties:
        _id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string }
        department: { type: string }
        must_change_password: { type: boolean }
    LoginSuccessResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/AuthUser'
...
    LecturerPaper:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        code: { type: string }
        course_id: { type: string }
        lecturer_id: { type: string }
        semester:
          type: integer
          nullable: true
        total_classes: { type: integer }
        course_name:
          type: string
          nullable: true
        course_code:
          type: string
          nullable: true
        course_status: { type: string }
        is_course_inactive: { type: boolean }
        academic_year:
          type: string
          nullable: true
        total_enrolled_students: { type: integer }
        enrolled_academic_sessions:
          type: array
          items: { type: string }
        enrolled_academic_session:
          type: string
          nullable: true
        enrolled_academic_session_label:
          type: string
          nullable: true
      additionalProperties: true
    LecturerSessionStartResponse:
      type: object
      required: [session_id, paper]
      properties:
        session_id: { type: string }
        paper:
          $ref: '#/components/schemas/LecturerPaper'
        started_at:
          type: string
          nullable: true
    RecognitionMatch:
      type: object
      properties:
        user_id: { type: string }
        name: { type: string }
        similarity: { type: number }
        confidence: { type: number }
      additionalProperties: true
    LecturerSessionRecognitionResponse:
      type: object
      required: [new_matches, faces_detected, candidates_count, threshold]
      properties:
        new_matches:
          type: array
          items:
            $ref: '#/components/schemas/RecognitionMatch'
        faces_detected: { type: integer }
        total_recognized:
          type: integer
          nullable: true
        candidates_count: { type: integer }
        threshold: { type: number }
        best_similarity_seen:
          type: number
          nullable: true
        saved_folder:
          type: string
          nullable: true
        original_path:
          type: string
          nullable: true
        face_paths:
          type: array
          items: { type: string }
          nullable: true
...
    AdminStudent:
      type: object
      properties:
        _id: { type: string }
        user_id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        course_id: { type: string }
        reg_number: { type: string }
        roll_number: { type: string }
        academic_session: { type: string }
        current_semester:
          type: integer
          nullable: true
      additionalProperties: true
...
    PaginatedStudentsResponse:
      type: object
      required: [items, total, page, per_page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AdminStudent' }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
    AdminStudentsListResponse:
      oneOf:
        - type: array
          items: { $ref: '#/components/schemas/AdminStudent' }
        - $ref: '#/components/schemas/PaginatedStudentsResponse'
...
    AdminCreateStudentResponse:
      type: object
      required: [temp_password]
      properties:
        temp_password: { type: string }
        profile:
          allOf:
            - $ref: '#/components/schemas/GenericObject'
          nullable: true
      additionalProperties: true
...
    AdminBulkPromoteResponse:
      type: object
      required: [message, promoted_count, skipped_count, skipped_max_semester_count, removed_papers_count]
      properties:
        message: { type: string }
        promoted_count: { type: integer }
        skipped_count: { type: integer }
        skipped_max_semester_count: { type: integer }
        removed_papers_count: { type: integer }
...
    AdminTrainingQueuedResponse:
      type: object
      required: [message, job_id, status_url, requested_count]
      properties:
        message: { type: string }
        job_id: { type: string }
        status_url: { type: string }
        requested_count: { type: integer }
...
    AdminTrainBulkResponse:
      type: object
      required: [message, requested_count, success_count, failure_count, total_trained_embeddings, items]
      properties:
        message: { type: string }
        requested_count: { type: integer }
        success_count: { type: integer }
        failure_count: { type: integer }
        total_trained_embeddings: { type: integer }
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        trainer:
          allOf:
            - $ref: '#/components/schemas/GenericObject'
          nullable: true
...
    AttendanceMatrixResponse:
      type: object
      required: [filters, options, meta, dates, rows]
      properties:
        filters:
          type: object
          additionalProperties: true
        options:
          type: object
          additionalProperties: true
        meta:
          type: object
          additionalProperties: true
        dates:
          type: array
          items:
            type: object
            additionalProperties: true
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
...
security:
  - jwtCookie: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: Logged in, content: { application/json: { schema: { $ref: '#/components/schemas/LoginSuccessResponse' } } } }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '401': { description: Invalid credentials, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '429': { description: Rate limited/locked, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
  /api/admin/papers/bulk-assign:
    post:
      tags: [AdminPapers]
      operationId: adminBulkAssignPaper
      security:
        - jwtCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                paper_id: { type: string }
                paper_ids: { type: array, items: { type: string } }
                lecturer_id: { type: string }
                course_id: { type: string }
                student_ids: { type: array, items: { type: string } }
      responses:
        '200': { description: Assignment completed, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } }
...
  /api/admin/students:
    get:
      tags: [AdminStudents]
      operationId: adminListStudents
      responses:
        '200': { description: Students list, content: { application/json: { schema: { $ref: '#/components/schemas/AdminStudentsListResponse' } } } }
    post:
      tags: [AdminStudents]
      operationId: adminCreateStudent
      security:
        - jwtCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStudentRequest' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/AdminCreateStudentResponse' } } } }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: Conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
  /api/admin/students/bulk-promote:
    post:
      tags: [AdminStudents]
      operationId: adminBulkPromoteStudents
      security:
        - jwtCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BulkPromoteRequest' }
      responses:
        '200': { description: Promotion summary, content: { application/json: { schema: { $ref: '#/components/schemas/AdminBulkPromoteResponse' } } } }
...
  /api/admin/students/train-face/bulk:
    post:
      tags: [AdminStudents]
      operationId: adminBulkTrainFace
      security:
        - jwtCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [student_ids]
              properties:
                student_ids: { type: array, items: { type: string } }
                async: { type: boolean }
      responses:
        '200': { description: Inline bulk training result, content: { application/json: { schema: { $ref: '#/components/schemas/AdminTrainBulkResponse' } } } }
        '202': { description: Queued bulk training result, content: { application/json: { schema: { $ref: '#/components/schemas/AdminTrainingQueuedResponse' } } } }
...
  /api/admin/attendance-matrix:
    get:
      tags: [AdminAttendance]
      operationId: adminAttendanceMatrix
      parameters:
        - in: query
          name: course_id
          required: true
          schema: { type: string }
        - in: query
          name: academic_session
          required: true
          schema: { type: string }
        - in: query
          name: semester
          required: true
          schema: { type: integer }
        - in: query
          name: from_date
          schema: { type: string }
        - in: query
          name: to_date
          schema: { type: string }
        - in: query
          name: tz_offset_minutes
          schema: { type: integer }
      responses:
        '200': { description: Attendance matrix, content: { application/json: { schema: { $ref: '#/components/schemas/AttendanceMatrixResponse' } } } }
...
  /api/lecturer/session/start:
    post:
      tags: [Lecturer]
      operationId: lecturerSessionStart
      security:
        - jwtCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartSessionRequest' }
      responses:
        '200': { description: Started, content: { application/json: { schema: { $ref: '#/components/schemas/LecturerSessionStartResponse' } } } }
  /api/lecturer/session/recognize:
    post:
      tags: [Lecturer]
      operationId: lecturerSessionRecognizeFrame
      security:
        - jwtCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecognizeFrameRequest' }
      responses:
        '200': { description: Recognition result, content: { application/json: { schema: { $ref: '#/components/schemas/LecturerSessionRecognitionResponse' } } } }
  /api/lecturer/session/recognize-image:
    post:
      tags: [Lecturer]
      operationId: lecturerSessionRecognizeImage
      security:
        - jwtCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [session_id, image]
              properties:
                session_id: { type: string }
                image: { type: string, format: binary }
      responses:
        '200': { description: Recognition result, content: { application/json: { schema: { $ref: '#/components/schemas/LecturerSessionRecognitionResponse' } } } }
...
  /api/lecturer/session/stop:
    post:
      tags: [Lecturer]
      operationId: lecturerSessionStop
      security:
        - jwtCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id]
              properties:
                session_id: { type: string }
      responses:
        '200': { description: Stopped, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } }
...